# unicorn-api/docs/swagger.yaml
basePath: /
definitions:
...
  handlers.CreateBucketRequest:
    description: Request to create a new storage bucket
    properties:
      name:
        description: |-
          The name of the bucket
          example: my-bucket
        type: string
    required:
    - name
    type: object
...
  handlers.ErrorResponse:
    properties:
      details:
        description: |-
          Additional error details
          example: Field 'email' is required
        type: string
      error:
        description: |-
          The error message
          example: Invalid request
        type: string
      status_code:
        description: |-
          The HTTP status code
          example: 400
        type: integer
      timestamp:
        description: |-
          The timestamp when the error occurred
          example: 2024-01-01T12:00:00Z
        type: string
    type: object
...
  models.File:
    description: A file stored in a storage bucket.
    properties:
      content_type:
        description: The MIME type of the file
        type: string
      contents:
        description: The file contents (base64 or text)
        type: string
      created_at:
        description: The creation timestamp
        type: string
      id:
        description: The unique identifier of the file
        type: string
      name:
        description: The name of the file
        type: string
      size:
        description: The size of the file in bytes
        type: integer
      storage_bucket_id:
        description: The ID of the bucket this file belongs to
        type: string
      updated_at:
        description: The last update timestamp
        type: string
    type: object
...
  models.StorageBucket:
    description: A storage bucket owned by a user, containing files.
    properties:
      created_at:
        description: The creation timestamp
        type: string
      files:
        description: The files in the bucket
        items:
          $ref: '#/definitions/models.File'
        type: array
      id:
        description: The unique identifier of the bucket
        type: string
      name:
        description: The name of the bucket
        type: string
      updated_at:
        description: The last update timestamp
        type: string
      user_id:
        description: The ID of the user who owns the bucket
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@unicorn-api.com
    name: Unicorn API Support
    url: https://github.com/your-org/unicorn-api
  description: A comprehensive RESTful API for Unicorn services providing Identity
    and Access Management (IAM) functionality. This API supports user authentication,
    role-based access control, organization management, and JWT token handling.
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Unicorn API
  version: "1.0"
paths:
  /api/v1/buckets:
    get:
      description: List all storage buckets owned by the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.StorageBucket'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden - insufficient permissions
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List buckets
      tags:
      - Storage
    post:
      consumes:
      - application/json
      description: Create a new storage bucket
      parameters:
      - description: Bucket name
        in: body
        name: bucket
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateBucketRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.StorageBucket'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden - insufficient permissions
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create bucket
      tags:
      - Storage
  /api/v1/buckets/{bucket_id}/files:
    get:
      description: List all files in a storage bucket
      parameters:
      - description: Bucket ID
        in: path
        name: bucket_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.File'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden - insufficient permissions
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List files
      tags:
      - Storage
    post:
      consumes:
      - multipart/form-data
      description: Upload a file to a storage bucket
      parameters:
      - description: Bucket ID
        in: path
        name: bucket_id
        required: true
        type: string
      - description: File to upload
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.File'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden - insufficient permissions
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Upload file
      tags:
      - Storage
  /api/v1/buckets/{bucket_id}/files/{file_id}:
    delete:
      description: Delete a file from a storage bucket
      parameters:
      - description: Bucket ID
        in: path
        name: bucket_id
        required: true
        type: string
      - description: File ID
        in: path
        name: file_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden - insufficient permissions
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete file
      tags:
      - Storage
    get:
      description: Download a file from a storage bucket
      parameters:
      - description: Bucket ID
        in: path
        name: bucket_id
        required: true
        type: string
      - description: File ID
        in: path
        name: file_id
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden - insufficient permissions
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Download file
      tags:
      - Storage
...
securityDefinitions:
  BearerAuth:
    description: 'Type "Bearer" followed by a space and JWT token. Example: "Bearer
      eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"